feat(fullstack): Long ID 전환 및 인증/피드 UI 정리

+++ back/src/main/kotlin/com/back/boundedContexts/member/adapter/persistence/MemberAttrRepository.kt
@@ -9,5 +9,5 @@ import org.springframework.data.jpa.repository.JpaRepository
  * - 주의: 변경 시 호출 경계와 데이터 흐름 영향을 함께 검토합니다.
  */
 interface MemberAttrRepository :
-    JpaRepository<MemberAttr, Int>,
+    JpaRepository<MemberAttr, Long>,
     MemberAttrRepositoryCustom
