+++ back/src/main/kotlin/com/back/boundedContexts/member/adapter/persistence/MemberRepository.kt
@@ -11,7 +11,7 @@ import org.springframework.data.jpa.repository.JpaRepository
  * - 주의: 변경 시 호출 경계와 데이터 흐름 영향을 함께 검토합니다.
  */
 interface MemberRepository :
-    JpaRepository<Member, Int>,
+    JpaRepository<Member, Long>,
     MemberRepositoryCustom {
     fun existsByEmail(email: String): Boolean
