+++ back/src/main/kotlin/com/back/boundedContexts/member/application/port/output/MemberAttrRepositoryPort.kt
@@ -30,7 +30,7 @@ interface MemberAttrRepositoryPort {
     ): Boolean
 
     fun existsBySubjectIdAndNameAndStrValueContaining(
-        subjectId: Int,
+        subjectId: Long,
         name: String,
         valueFragment: String,
     ): Boolean
