+++ back/src/main/kotlin/com/back/boundedContexts/member/subContexts/signupVerification/adapter/persistence/MemberSignupVerificationRepository.kt
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository
  * - 역할: 계층 간 계약(포트/스펙) 정의를 담당합니다.
  * - 주의: 변경 시 호출 경계와 데이터 흐름 영향을 함께 검토합니다.
  */
-interface MemberSignupVerificationRepository : JpaRepository<MemberSignupVerification, Int> {
+interface MemberSignupVerificationRepository : JpaRepository<MemberSignupVerification, Long> {
     fun findByEmailVerificationToken(emailVerificationToken: String): MemberSignupVerification?
 
     fun findBySignupSessionToken(signupSessionToken: String): MemberSignupVerification?
