+++ back/src/main/kotlin/com/back/boundedContexts/post/adapter/persistence/PostAttrRepositoryImpl.kt
@@ -91,7 +91,7 @@ class PostAttrRepositoryImpl : PostAttrRepositoryCustom {
      * IntValue 항목을 수정한다.
      */
     private fun updateIntValue(
-        subjectId: Int,
+        subjectId: Long,
         name: String,
         delta: Int,
     ): List<Int> =
